requirements.txt
boto3
logging
requests
PyYAML